requirements.txt
Django==4.2.20
Pillow==11.1.0

gunicorn>=20.0.0
django-allauth==0.63.6
django-debug-toolbar==3.6.0
django-environ==0.9.0
django-extensions==3.2.1
djangorestframework==3.15.2

django-storages~=1.14.6
boto3
python-decouple
djangorestframework-simplejwt
drf-spectacular
django-debug-toolbar

flake8==5.0.4
isort==5.10.1

stripe==4.1.0
importlib-metadata==4.13.0

redis==4.5.4
celery==5.2.7
django-redis==5.2.0

psycopg2-binary==2.9.10